# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Identity Platform - Authentication & Authorization Server
  version: 1.0.0
  description: |
    Centralized authentication and authorization platform built on Express.

    Features: Argon2id passwords, EdDSA (Ed25519) JWTs with JWKS publishing,
    OAuth2 Authorization Code + PKCE and client_credentials grants, refresh-token
    rotation with theft detection, TOTP MFA, RBAC with permission-scoped admin API,
    Redis-backed rate limiting, risk engine, audit logging, RFC 7662 introspection
    and RFC 7009 revocation.

    Access tokens are EdDSA-signed JWTs; verify them against `/.well-known/jwks.json`.
servers:
  - url: http://localhost:3001
    description: Auth server (direct)
  - url: http://localhost:3000
    description: API gateway (resource traffic)
...
paths:
...
  /admin/users:
    get:
      tags: [Admin]
      summary: List users (users:read)
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { in: query, name: offset, schema: { type: integer, minimum: 0 } }
      responses:
        '200': { description: User list with roles }
        '403': { description: MISSING_PERMISSION }
  /admin/users/{id}/roles:
    put:
      tags: [Admin]
      summary: Replace a user's roles (users:write)
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roles]
              properties:
                roles: { type: array, items: { type: string } }
      responses:
        '200': { description: Roles updated }
  /admin/users/{id}/status:
    put:
      tags: [Admin]
      summary: Suspend/reactivate a user (users:write)
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: Status updated }
  /admin/roles:
    get:
      tags: [Admin]
      summary: List roles (roles:read)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Role list }
  /admin/roles/{id}/permissions:
    put:
      tags: [Admin]
      summary: Replace a role's permissions (roles:write)
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: Permissions updated }
...
  /admin/audit:
    get:
      tags: [Admin]
      summary: Query audit events (audit:read)
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: event_type, schema: { type: string } }
        - { in: query, name: limit, schema: { type: integer, maximum: 200 } }
      responses:
        '200': { description: Audit events }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: EdJWT
      description: EdDSA (Ed25519) signed access token, aud=api
    basicAuth:
      type: http
      scheme: basic
      description: OAuth client credentials (client_secret_basic)
...